update CSS 02/05/23
added the scissors hand and box. positioned it to the left of the screen.

changed all of the measurements for the height and width of every hand to percentages so that they would fit any screen.
unfortunately position of the boxes is still wonky, need to figure out how to better align those for more aesthetically pleasing placement for viewers.