grid_perc_update.css
Switched to a grid layout, rather than using left and right orientation values. As well as swapped out the height and width's back to percentiles.

issue with the previous, where viewport width and height were input, was that the boxes would often be shaped like a long rectangle and distort the hand signs within. by swapping to percentiles they retain a more square shape which is what I was aiming for in the first place.