Update 03.07.23 Gradient BG & Single Box
Got rid of the 3 boxes with each hand symbol. And instead opted to go with a single box where all of them will cycle through.
This makes the display a lot simpler and easier on the eyes as the alignment of all of them has been fighting me the whole time.

Also changed the background from a plane white to the colors of the symbols. May turn them darker in the future.