Update r-p-s-Shoot 03.12.23
cleared out everything previously on the js document. Instead created the array with the string of each div for the symbols.
then created a function called 'shoot()' to run with the html onClick function.
And lastly: started on a function for the animation, which will eventually be the primary function of the html button, of all the hand symbols that will then call on the shoot() function.